cluster: fix odr violation in rm_stm compat test
47: ==648522==ERROR: AddressSanitizer: odr-violation (0x5611f89d4220):
47:   [1] size=240 'vtable for cluster::rm_stm' /home/nwatkins/src/redpanda/src/v/cluster/tests/rm_stm_compatibility_test.cc
47:   [2] size=240 'vtable for cluster::rm_stm' /home/nwatkins/src/redpanda/src/v/cluster/rm_stm.cc
