+++ src/v/cluster/rm_stm.cc
@@ -113,13 +113,6 @@ model::record_batch make_fence_batch_v2(
     return std::move(builder).build();
 }
 
-struct fence_batch_data {
-    model::batch_identity bid;
-    std::optional<model::tx_seq> tx_seq;
-    std::optional<std::chrono::milliseconds> transaction_timeout_ms;
-    model::partition_id tm;
-};
-
 fence_batch_data read_fence_batch(model::record_batch&& b) {
     const auto& hdr = b.header();
     auto bid = model::batch_identity::from(hdr);
